+++ frontend/utils/shop_util.js
@@ -1,3 +1,5 @@
+import {debug} from "./tools";
+
 export const fetchShops = () => {
     return $.ajax({
         url: '/api/shops',
@@ -6,6 +8,10 @@ export const fetchShops = () => {
 };
 
 export const fetchShop = (shopId) => {
+    if (!shopId) {
+        return debug.error(`A shop id must be provided for fetchShop.`)
+    }
+
     return $.ajax({
         url: `/api/shops/${shopId}`,
         method: 'GET'
@@ -13,6 +19,10 @@ export const fetchShop = (shopId) => {
 };
 
 export const createShop = (shop) => {
+    if (!shop) {
+        return debug.error(`A shop object must be provided for createShop.`)
+    }
+
     return $.ajax({
         url: '/api/shops',
         method: 'POST',
@@ -21,6 +31,10 @@ export const createShop = (shop) => {
 };
 
 export const updateShop = (shop) => {
+    if (!shop) {
+        return debug.error(`A shop object must be provided for updateShop.`)
+    }
+
     return $.ajax({
         url: `/api/shops/${shop.id}`,
         method: 'PATCH',
@@ -29,6 +43,10 @@ export const updateShop = (shop) => {
 };
 
 export const deleteShop = (shopId) => {
+    if (!shopId) {
+        return debug.error(`A shop id must be provided for deleteShop.`)
+    }
+
     return $.ajax({
         url: `/api/shops/${shopId}`,
         method: 'DELETE'
